можно удалять краточку

+++ src/utils/Api.js
@@ -66,36 +66,16 @@ class Api {
 
     // удаление карточки
 
-    deleteCard(cardId) {
-        return fetch(`${this._adress}/cards/${cardId}`, {
-            method: "DELETE",
-            headers: {
-                authorization: this._token,
-            }
-        }) .then(this._getResponseData)
+    deleteCard(cardId, isMine) {
+        if(isMine) {
+                return fetch(`${this._adress}/cards/${cardId}`, {
+                method: "DELETE",
+                headers: {
+                    authorization: this._token,
+                }
+            }) .then(this._getResponseData)
+        }
     }
-
-    // добавление лайка
-
-    // addLike(cardId) {
-    //     return fetch(`${this._adress}/cards/${cardId}/likes`, {
-    //         method: "PUT",
-    //         headers: {
-    //             authorization: this._token,
-    //         }
-    //     }) .then(this._getResponseData)
-    // }
-
-    // // удаление лайка
-
-    // deleteLike(cardId) {
-    //     return fetch(`${this._adress}/cards/${cardId}/likes`, {
-    //         method: "DELETE",
-    //         headers: {
-    //             authorization: this._token,
-    //         }
-    //     }) .then(this._getResponseData)
-    // }
 
     // изменение лайка (добавление или удаление)
     changeLikeCardStatus(cardId, isLiked) {
